+++ run_yaml_indent.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+
+from yaml_indent.yaml_indent import main
+
+if __name__ == "__main__":
+    main()