+++ yaml_indent/yaml_indent.py
@@ -7,10 +7,24 @@ from ruamel.yaml import YAML
 from . import __version__
 
 def find_config_file(dir_path):
-    '''Finds config file for files in dir_path'''
+    '''
+    Searches for the configuration file (.yaml_indent.ini) in the
+    given directory and its parent directories up to the home
+    directory.
+
+    Arguments:
+    - dir_path (str): The directory path where to start the search for the configuration file.
+
+    Returns:
+    - str: The path to the configuration file if found, None otherwise.
+
+    The function starts from the directory specified by dir_path and moves up towards the home directory.
+    The search stops when it finds a configuration file or reaches the home directory.
+    If the function cannot find a configuration file, it returns None.
+
+    '''
     home_dir = os.path.expanduser("~")  # Home directory path
-    while dir_path != home_dir and dir_path != "/" and dir_path != "":  # Stop at the home directory
-        print(dir_path)
+    while dir_path not in {home_dir, "/", ""}:  # Stop at the home directory
         config_path = os.path.join(dir_path, '.yaml_indent.ini')
         if os.path.exists(config_path):
             return config_path
@@ -20,11 +34,26 @@ def find_config_file(dir_path):
 
 
 def process_yaml_file(input_file, output_file=None, in_place=False):
-    '''Indent input_file according to default rules or rules found in the
-    config_file'''
+    '''
+    Reindents the YAML file(s) based on the specified configuration or default rules.
+    
+    Arguments:
+    - input_file (str): The path to the input YAML file.
+    - output_file (str, optional): The path to the output YAML file. If not provided, the output is printed to stdout.
+    - in_place (bool, optional): If true, modifies the input file in place. Default is False.
 
+    The function reads the configuration from a file found in the current directory or any parent directory,
+    and applies it to reindent the input YAML file. The configuration file should be named '.yaml_indent.ini'.
+    If no configuration file is found, default values are used.
+
+    The default values for indentation are:
+    - mapping: 2
+    - sequence: 4
+    - offset: 2
+    '''
     yaml = YAML()
-
+    yaml.explicit_start = True  # Start each document with '---'
+    
     # Check for a configuration file
     config_path = find_config_file(os.path.dirname(input_file))
     if config_path:
@@ -42,29 +71,32 @@ def process_yaml_file(input_file, output_file=None, in_place=False):
     # Open the input YAML file and load the data
     with open(input_file, 'r', encoding='UTF-8') as stream:
         try:
-            data = yaml.load(stream)
+            documents = list(yaml.load_all(stream))
         except Exception as exc:
-            print(exc)
+            print(f"Error loading YAML data from {input_file}: {exc}")
+            return
 
     # Open the output file and dump the data with correct indenting
     if in_place:
         with open(input_file, 'w', encoding='UTF-8') as outfile:
             try:
-                yaml.dump(data, outfile)
+                for doc in documents:
+                    yaml.dump(doc, outfile)
             except Exception as exc:
-                print(exc)
+                print(f"Error writing YAML data to {input_file}: {exc}")
     elif output_file:
         with open(output_file, 'w', encoding='UTF-8') as outfile:
             try:
-                yaml.dump(data, outfile)
+                for doc in documents:
+                    yaml.dump(doc, outfile)
             except Exception as exc:
-                print(exc)
+                print(f"Error writing YAML data to {output_file}: {exc}")
     else:
         try:
-            yaml.dump(data, sys.stdout)
+            for doc in documents:
+                yaml.dump(doc, sys.stdout)
         except Exception as exc:
-            print(exc)
-
+            print(f"Error writing YAML data to stdout: {exc}")
 
 def main():
     '''Parse arguments and indent files'''